refactor: abstracted new NoteCard component from DisplayGrid
