+++ src/components/DisplayGrid/display_grid.css
@@ -4,47 +4,3 @@
   grid-template-columns: repeat(3, minmax(0, 1fr));
   grid-gap: 2rem;
 }
-
-.grid-item {
-  padding: 1rem;
-  border: 0.1rem solid var(--text-color-primary);
-  word-wrap: break-word;
-  border-bottom-right-radius: 1rem;
-  position: relative;
-}
-
-.item-btn {
-  border-radius: 50%;
-  border: 0.1rem solid var(--text-color-primary);
-  background-color: var(--bg-color-primary);
-  color: var(--text-color-primary);
-  height: 1.5rem;
-  width: 1.5rem;
-  font-weight: 700;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-}
-
-.close {
-  position: absolute;
-  right: -0.6rem;
-  top: -0.8rem;
-}
-
-.date {
-  position: absolute;
-  right: 1.5rem;
-  top: -0.65rem;
-  padding: 0 0.2rem;
-  border: 0.1rem solid var(--text-color-primary);
-  background-color: var(--bg-color-primary);
-}
-
-.user-msg {
-  font-size: 2rem;
-  grid-column: span 3;
-  justify-self: center;
-  text-align: center;
-}
