+++ src/components/index.js
@@ -1,3 +1,4 @@
 export {Header} from "./Header/Header.jsx";
 export {Form} from "./Form/Form.jsx";
 export {DisplayGrid} from "./DisplayGrid/DisplayGrid.jsx";
+export {NoteCard} from "./NoteCard/NoteCard";
